refactor(hooks): extract language prefix logic into custom hook
Move the language prefix logic from LanguageLink component into a reusable custom hook (useLanguagePrefix) to improve code maintainability and reusability. This change also includes updates to tsconfig.app.json and vite.config.ts to support the new hooks directory.
