+++ vite.config.ts
@@ -18,6 +18,7 @@ export default defineConfig({
       "@store": path.resolve(__dirname, "./src/store"),
       "@ctypes": path.resolve(__dirname, "./src/types"),
       "@config": path.resolve(__dirname, "./src/config"),
+      "@hooks": path.resolve(__dirname, "./src/hooks"),
     },
   },
 });
